Split syntax module between text and binary
