+++ dhall/src/syntax/binary/mod.rs
@@ -1,0 +1,2 @@
+pub mod binary;
+pub(crate) use binary::{encode, decode};
