+++ dhall/src/syntax/mod.rs
@@ -9,8 +9,7 @@ mod core;
 pub use crate::syntax::core::context;
 pub use crate::syntax::core::visitor;
 pub use crate::syntax::core::*;
-mod printer;
-pub use crate::syntax::printer::*;
-mod parser;
-pub use crate::syntax::parser::*;
+pub use crate::syntax::text::printer::*;
+pub use crate::syntax::text::parser::*;
 pub mod binary;
+pub mod text;
